requirements.txt
torch==2.1.0
torchdata==0.7.0
torchdiffeq==0.2.3
numpy==1.24.4
altair==5.4.1
rdkit-pypi==2022.9.5
pandas>1.0
jupyter==1.1.1
matplotlib==3.9.2
scipy==1.14.1
scikit-learn==1.5.1
boto3==1.35.7
botocore==1.35.7
tqdm==4.66.5
torchinfo==1.8.0
einops==0.8.0
smart-open==7.0.4
seaborn==0.13.2
zuko==1.2.0
s3fs==2024.6.0
due @ git+https://github.com/y0ast/DUE.git